third_party/WebKit/LayoutTests/fast/alignment/parse-justify-self-expected.txt

Test that setting and getting justify-self works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting justify-self set through CSS
PASS getComputedStyle(justifySelfBaseline, '').getPropertyValue('justify-self') is 'baseline'
PASS getComputedStyle(justifySelfLastBaseline, '').getPropertyValue('justify-self') is 'last-baseline'
PASS getComputedStyle(justifySelfStretch, '').getPropertyValue('justify-self') is 'stretch'
PASS getComputedStyle(justifySelfStart, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(justifySelfEnd, '').getPropertyValue('justify-self') is 'end'
PASS getComputedStyle(justifySelfCenter, '').getPropertyValue('justify-self') is 'center'
PASS getComputedStyle(justifySelfSelfEnd, '').getPropertyValue('justify-self') is 'self-end'
PASS getComputedStyle(justifySelfSelfStart, '').getPropertyValue('justify-self') is 'self-start'
PASS getComputedStyle(justifySelfLeft, '').getPropertyValue('justify-self') is 'left'
PASS getComputedStyle(justifySelfRight, '').getPropertyValue('justify-self') is 'right'
PASS getComputedStyle(justifySelfFlexStart, '').getPropertyValue('justify-self') is 'flex-start'
PASS getComputedStyle(justifySelfFlexEnd, '').getPropertyValue('justify-self') is 'flex-end'
PASS getComputedStyle(justifySelfEndTrue, '').getPropertyValue('justify-self') is 'end true'
PASS getComputedStyle(justifySelfCenterTrue, '').getPropertyValue('justify-self') is 'center true'
PASS getComputedStyle(justifySelfSelfEndSafe, '').getPropertyValue('justify-self') is 'self-end safe'
PASS getComputedStyle(justifySelfSelfStartSafe, '').getPropertyValue('justify-self') is 'self-start safe'
PASS getComputedStyle(justifySelfRightSafe, '').getPropertyValue('justify-self') is 'right safe'
PASS getComputedStyle(justifySelfLeftTrue, '').getPropertyValue('justify-self') is 'left true'
PASS getComputedStyle(justifySelfFlexStartTrue, '').getPropertyValue('justify-self') is 'flex-start true'
PASS getComputedStyle(justifySelfFlexEndSafe, '').getPropertyValue('justify-self') is 'flex-end safe'

Test initial value of justify-self through JS
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'

Test getting and setting justify-self through JS
PASS element.style.justifySelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "center"
PASS element.style.justifySelf is "start true"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start true"
PASS element.style.justifySelf is "flex-end safe"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "flex-end safe"
PASS element.style.justifySelf is "right"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "right"
PASS element.style.justifySelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "center"
PASS element.style.justifySelf is "self-start"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "self-start"
PASS element.style.justifySelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"
PASS element.style.justifySelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"
PASS element.style.justifySelf is "self-end"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "self-end"

Test bad combinations of justify-self
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"
PASS element.style.justifySelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"

Test the value 'initial'
PASS element.style.justifySelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "center"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "start"

Test the value 'initial' for grid containers
PASS element.style.justifySelf is "left safe"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "left safe"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"

Test the value 'initial' for flex containers
PASS element.style.justifySelf is "right true"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "right true"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"

Test the value 'initial' for positioned elements
PASS element.style.justifySelf is "left"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "left"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"

Test the value 'initial' for positioned elements in grid containers
PASS element.style.justifySelf is "right"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "right"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"

Test the value 'initial' for positioned elements in grid containers
PASS element.style.justifySelf is "end"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "end"
PASS element.style.justifySelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "stretch"

Test the value 'inherit'
PASS element.style.justifySelf is "end"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "end"
PASS element.style.justifySelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "end"
PASS element.style.justifySelf is "left safe"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "left safe"
PASS element.style.justifySelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "left safe"
PASS element.style.justifySelf is "center true"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "center true"
PASS element.style.justifySelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('justify-self') is "center true"
PASS successfullyParsed is true

TEST COMPLETE
